requirements.txt
fastapi>=0.128.0
uvicorn>=0.30.0
groq>=0.33.0
langchain>=1.2.0
langchain-groq>=0.1.3
langchain-community>=0.4.0
langchain-core>=1.2.0
langchain-text-splitters>=1.1.0
chromadb>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
jinja2>=3.1.0
python-multipart>=0.0.6
requests>=2.31.0
pdfplumber>=0.11.0
python-docx>=1.1.0
